repalce manifest service search usage with EmanifestSearch class usage
