+++ server/apps/manifest/services/emanifest_search.py
@@ -1,4 +1,4 @@
-from datetime import datetime, timezone
+from datetime import UTC, datetime, timedelta, timezone
 from typing import Literal, Optional, get_args
 
 from apps.core.services import RcrainfoService
@@ -71,6 +71,28 @@ class EmanifestSearch:
     def _emanifest_correction_request_status(cls, correction_request_status) -> bool:
         return cls.__validate_literal(correction_request_status, CorrectionRequestStatus)
 
+    def _date_or_three_years_past(self, start_date: Optional[datetime]) -> str:
+        return (
+            start_date.replace(tzinfo=timezone.utc).strftime(self.rcra_client.datetime_format)
+            if start_date
+            else (
+                datetime.now(UTC)
+                - timedelta(
+                    minutes=60  # 60 seconds/1minutes
+                    * 24  # 24 hours/1day
+                    * 30  # 30 days/1month
+                    * 36  # 36 months/3years = 3/years
+                )
+            ).strftime(self.rcra_client.datetime_format)
+        )
+
+    def _date_or_now(self, end_date: Optional[datetime]) -> str:
+        return (
+            end_date.replace(tzinfo=timezone.utc).strftime(self.rcra_client.datetime_format)
+            if end_date
+            else datetime.now(UTC).strftime(self.rcra_client.datetime_format)
+        )
+
     def build_search_args(self):
         search_params = {
             "stateCode": self.state_code,
@@ -120,13 +142,11 @@ class EmanifestSearch:
         return self
 
     def add_start_date(self, start_date: datetime = None):
-        self.start_date = start_date
+        self.start_date = self._date_or_three_years_past(start_date)
         return self
 
     def add_end_date(self, end_date: datetime = None):
-        if not end_date:
-            end_date = datetime.now().replace(tzinfo=timezone.utc)
-        self.end_date = end_date
+        self.end_date = self._date_or_now(end_date)
         return self
 
     def execute(self):
